chore: move atomic folder to internal/ dir
