+++ internal/atomic/counter.go
